Update README.md
I had a follow-up task to determine the time it takes for a user to make a purchase on the website. I was tasked with finding the duration from a user's initial visit to the website on any given day until their first purchase on that same day. My final result displays the dynamic daily duration.